backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
PyJWT==2.3.0
bcrypt==3.2.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pymongo==4.6.1 